+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added [Get-YamlFromMarkdown][] function.
 - Added [Infer-ObjectSchema][] function.
+- Added [Get-Dataset][] `-Library` option.
 
 ## [0.11.0] - 2025-06-25